refactor: Rename buildPaperslistByContact to buildFilesByContacts
Move buildFilesByContacts to Parpers/helpers.js
