+++ packages/cozy-mespapiers-lib/src/components/Papers/helpers.spec.js
@@ -1,4 +1,8 @@
-import { harmonizeContactsNames, groupFilesByContacts } from './helpers'
+import {
+  harmonizeContactsNames,
+  groupFilesByContacts,
+  buildFilesByContacts
+} from './helpers'
 
 const mockContacts00 = [
   { _id: 'contactId01', name: { givenName: 'Bob', familyName: 'Durand' } },
@@ -47,6 +51,17 @@ const mockFiles = [
   {
     _id: 'fileId04',
     name: 'file04.pdf'
+  }
+]
+
+const mockFilesWithoutContact = [
+  {
+    _id: 'fileId01',
+    name: 'file01.pdf'
+  },
+  {
+    _id: 'fileId02',
+    name: 'file02.pdf'
   }
 ]
 
@@ -113,6 +128,123 @@ describe('helpers Papers', () => {
       const res = groupFilesByContacts(mockFiles, mockContacts00)
 
       expect(res).toStrictEqual(expected)
+    })
+  })
+
+  describe('buildFilesByContacts', () => {
+    it('should return object with all papers filtered by contacts', () => {
+      const expected = [
+        {
+          withHeader: true,
+          contact: 'Alice Durand',
+          papers: {
+            maxDisplay: 3,
+            list: [
+              {
+                _id: 'fileId02',
+                name: 'file02.pdf',
+                relationships: {
+                  referenced_by: {
+                    data: [{ id: 'contactId02', type: 'io.cozy.contacts' }]
+                  }
+                }
+              }
+            ]
+          }
+        },
+        {
+          withHeader: true,
+          contact: 'Bob Durand',
+          papers: {
+            maxDisplay: 3,
+            list: [
+              {
+                _id: 'fileId01',
+                name: 'file01.pdf',
+                relationships: {
+                  referenced_by: {
+                    data: [{ id: 'contactId01', type: 'io.cozy.contacts' }]
+                  }
+                }
+              }
+            ]
+          }
+        },
+        {
+          withHeader: true,
+          contact: 'PapersList.contactMerged',
+          papers: {
+            maxDisplay: 3,
+            list: [
+              {
+                _id: 'fileId03',
+                name: 'file03.pdf',
+                relationships: {
+                  referenced_by: {
+                    data: [
+                      { id: 'contactId01', type: 'io.cozy.contacts' },
+                      { id: 'contactId02', type: 'io.cozy.contacts' }
+                    ]
+                  }
+                }
+              }
+            ]
+          }
+        },
+        {
+          withHeader: true,
+          contact: 'PapersList.defaultName',
+          papers: {
+            maxDisplay: 3,
+            list: [
+              {
+                _id: 'fileId04',
+                name: 'file04.pdf'
+              }
+            ]
+          }
+        }
+      ]
+
+      const result = buildFilesByContacts({
+        files: mockFiles,
+        contacts: mockContacts00,
+        maxDisplay: 3,
+        t: jest.fn(key => key)
+      })
+
+      expect(result).toStrictEqual(expected)
+    })
+
+    it('should have not header if there are only files without contact', () => {
+      const expected = [
+        {
+          withHeader: false,
+          contact: 'PapersList.defaultName',
+          papers: {
+            maxDisplay: 3,
+            list: [
+              {
+                _id: 'fileId01',
+                name: 'file01.pdf'
+              },
+              {
+                _id: 'fileId02',
+                name: 'file02.pdf'
+              }
+            ]
+          }
+        }
+      ]
+
+      const result = buildFilesByContacts({
+        files: mockFilesWithoutContact,
+        contacts: mockContacts00,
+        maxDisplay: 3,
+        t: jest.fn(key => key)
+      })
+
+      expect(result).toStrictEqual(expected)
     })
   })
 })
